mqtt-to-db/requirements.txt
paho-mqtt==1.6.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
